+++ component/async/kafka/option_test.go
@@ -1,9 +1,7 @@
 package kafka
 
 import (
-	"errors"
 	"reflect"
-	"runtime"
 	"testing"
 	"time"
 
@@ -151,55 +149,4 @@ func TestDecoderJSON(t *testing.T) {
 		reflect.ValueOf(json.DecodeRaw).Pointer(),
 		reflect.ValueOf(c.DecoderFunc).Pointer(),
 	)
-}
-
-func TestWithDurationOffset(t *testing.T) {
-	f := func(_ *sarama.ConsumerMessage) (time.Time, error) {
-		return time.Time{}, nil
-	}
-
-	type args struct {
-		since         time.Duration
-		timeExtractor TimeExtractor
-	}
-	testCases := map[string]struct {
-		args        args
-		expectedErr error
-	}{
-		"success": {
-			args: args{
-				since:         time.Second,
-				timeExtractor: f,
-			},
-		},
-		"error - negative since duration": {
-			args: args{
-				since:         -time.Second,
-				timeExtractor: f,
-			},
-			expectedErr: errors.New("duration must be positive"),
-		},
-		"error - nil time extractor": {
-			args: args{
-				since: time.Second,
-			},
-			expectedErr: errors.New("empty time extractor function"),
-		},
-	}
-	for name, tt := range testCases {
-		t.Run(name, func(t *testing.T) {
-			c := ConsumerConfig{}
-			err := WithDurationOffset(tt.args.since, tt.args.timeExtractor)(&c)
-			if tt.expectedErr != nil {
-				assert.EqualError(t, err, tt.expectedErr.Error())
-			} else {
-				assert.NoError(t, err)
-				assert.True(t, c.DurationBasedConsumer)
-				assert.Equal(t, time.Second, c.DurationOffset)
-				assert.Equal(t,
-					runtime.FuncForPC(reflect.ValueOf(tt.args.timeExtractor).Pointer()).Name(),
-					runtime.FuncForPC(reflect.ValueOf(c.TimeExtractor).Pointer()).Name())
-			}
-		})
-	}
 }
